YetAnotherPlayground project with SPH code included from the old project.
This is a buildable version that has a lot of helper code and
learning/testing classes. All SPH code is in the same condition that it
was when I finished my master thesis. Only change is that instead of my
vec2d and vec3d classes for math computation I switched to GLM graphics
math library.